Moins de bugs

+++ C_Language/TP_SYS/tp_sema/sema.h
@@ -1,13 +1,3 @@
-#include <stdlib.h>
-#include <assert.h>
-#include <unistd.h>
-#include <fcntl.h>
-#include <stdio.h>
-#include <stdarg.h>
-#include <errno.h>
-#include <sys/types.h>
-#include <sys/ipc.h>
-#include <sys/sem.h>
 
 void erreur(char *s);
 int open_sem(key_t cle);
@@ -16,7 +6,7 @@ void V(int semid);
 
 /*-----------------------------------------------------------*/
 
-int creat_sem(key_t cle, ushort val);
+int creat_sem(key_t cle, int val);
 /* creation du semaphore cle ayant comme valeur initiale val
    retourne un identificateur de semaphore >=0 ou -1 si erreur
 */
